Remove HttpClient completely and fix home spec file name
