+++ src/app/courses/courses.component.spec.ts
@@ -3,7 +3,6 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { CoursesComponent } from './courses.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { CouchService } from '../shared/couchdb.service';
-// import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
 import { CourseValidatorService } from '../validators/course-validator.service';
 import { FormErrorMessagesComponent } from '../shared/form-error-messages.component';
