+++ src/app/home/navigation.component.spec.ts
@@ -3,7 +3,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 import { CommonModule } from '@angular/common';
-// import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
 import { NavigationComponent } from './navigation.component';
 import { CouchService } from '../shared/couchdb.service';
