+++ src/app/meetups/meetups-add.component.spec.ts
@@ -3,7 +3,6 @@ import { CouchService } from '../shared/couchdb.service';
 import { MeetupsAddComponent } from './meetups-add.component';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
-// import { HttpModule } from '@angular/http';
 
 describe('MeetupsAddComponent', () => {
