+++ src/app/validators/nation-validator.service.spec.ts
@@ -2,7 +2,6 @@ import { TestBed, inject } from '@angular/core/testing';
 import { CouchService } from '../shared/couchdb.service';
 import { HttpClientModule } from '@angular/common/http';
 import { NationValidatorService } from './nation-validator.service';
-// import { HttpModule } from '@angular/http';
 
 describe('NationValidatorService', () => {
   beforeEach(() => {
